+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This file is used to list changes made in each version of the openstack-block-st
 * Update gpfs volume driver path and fix gpfs_images_share_mode initialization
 * Change to cinder v2 service and endpoint
 * Add support for san_password with ibm.storwize_svc.StorwizeSVCDriver
+* Add glance_api_version config option
 
 ## 10.0.0
 * Upgrading to Juno